requirements.txt
absl-py==1.4.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
exceptiongroup==1.1.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.25.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyKakao==0.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rs==0.0.2
rsa==4.9
scipy==1.11.1
selenium==4.11.2
Selenium-Screenshot==1.7.0
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==4.0.0
Werkzeug==2.3.6
windows-curses==2.3.1
wsproto==1.2.0
zipp==3.16.2
